use simple redis lock: only one batcher instance
this kind of redis lock is simple because it's based on accessing one
redis instance, just like Batcher. So long as we set an N second delay
on our redis instance restart, mutual exclusion of Batcher clients is
guaranteed (N is the expire time on locks). If Batcher changes to depend
on a distributed Redis layer, the [redlock algorithm][1] should be used
instead.

[1]: https://redis.io/topics/distlock