limit concurrency / speed
An important aspect of batching is to decide how quickly to process
batches. For instance, a computationally expensive batch processing job
might have 5 worker machine process jobs. These 5 worker machines might
have other important work to do that we don't want to block with these
expensive jobs. We may wish to set a concurrency of 5, and a
minDelaySeconds to 60; now batcher will produce 5 batches per minute.
The best analogy is that there are 5 people preparing a batch, and each
batch takes at least a minute to prepare. Putting this functionality in
batcher itself, instead of consumers of the output stream, allows for 1)
"dumb" consumers that process batches as soon as they are made available
and 2) batches to be prepared on a realistic processing schedule